Trying to convert to sql values

+++ src/Parse.hs
@@ -21,7 +21,7 @@ data Currency = Currency {
     symbol :: String,
     rate :: String,
     description :: String,
-    rate_float :: Float
+    rate_float :: Double
 } deriving (Show, Generic)
 
 data Bpi = Bpi {
